// ClientApp/src/app/shared/models/shift-master.ts
export interface ShiftMasterModel {
    plantShiftId: string;
    plantId: string;
    lineId: string;
    plantName:string;
    shiftName:string;
    shiftId: string;
    startTime: string;
    hour:string;
    endTime:string;
    minute:string;
    timeZone: string;
    shiftDisplayTime:string;
    active: string;
}